# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sistema de Calificaciones - API",
    "version": "0.2.0",
    "description": "API REST del Sistema de Calificaciones Escolar. Multi-tenencia por escuela. Autenticación mediante cookie de sesión firmada (HMAC-SHA256) o Bearer token (para automatizaciones)."
  },
  "servers": [
    { "url": "/", "description": "Servidor actual" }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Cookie de sesión firmada con HMAC-SHA256."
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token para automatizaciones (AGENT_SECRET_TOKEN). Solo para endpoints específicos."
      }
    },
...
  },
  "security": [{ "sessionCookie": [] }],
  "paths": {
...
    "/api/escuelas": {
      "get": { "summary": "Listar escuelas", "security": [], "responses": { "200": { "description": "Lista de escuelas" } } },
      "post": {
        "summary": "Crear nueva escuela (clona estructura de escuela plantilla)",
        "security": [{ "sessionCookie": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nombre", "codigo", "adminEmail", "adminPassword", "adminNombre"],
                "properties": {
                  "nombre": { "type": "string" },
                  "codigo": { "type": "string" },
                  "cloneFromEscuelaId": { "type": "string", "description": "ID de la escuela plantilla a clonar (grados, materias, config). Default: primera escuela activa." },
                  "adminEmail": { "type": "string" },
                  "adminPassword": { "type": "string" },
                  "adminNombre": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Escuela creada con estructura clonada" }, "403": { "description": "Requiere superadmin" } }
      }
    },
...
  }
}